Update exercise "05. Shape".

+++ DevStyleR/C++ Programing Part 2/02. SecondLecture/Polymorphism/05. Shape/3DShape.cpp
@@ -2,7 +2,7 @@
 
 ThreeDShape::ThreeDShape(int a, int h) : Shape(a), h(h) {}
 
-int ThreeDShape::getArea() {
+double ThreeDShape::getArea() {
 	std::cout << "In 3D getArea." << std::endl;
 
 	return 0;
